@@ -4,13 +4,13 @@ version: 0.0.7
 homepage: https://decodedhealth.com
 issue_tracker: https://github.com/decodedhealth/flutter_zoom_plugin/issues
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0-0 <3.0.0'
   flutter: ">=1.12.0 <2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  archive: ^2.0.13
+  archive: ^3.1.2
 
 dev_dependencies:
   flutter_test:
