@@ -1,22 +1,20 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:flutter_zoom_plugin/zoom_view.dart';
-import 'package:flutter_zoom_plugin/zoom_options.dart';
-
 import 'package:flutter/material.dart';
+import 'package:flutter_zoom_plugin/zoom_options.dart';
+import 'package:flutter_zoom_plugin/zoom_view.dart';
 
 class MeetingWidget extends StatelessWidget {
+  ZoomOptions? zoomOptions;
+  ZoomMeetingOptions? meetingOptions;
 
-  ZoomOptions zoomOptions;
-  ZoomMeetingOptions meetingOptions;
-
-  Timer timer;
+  Timer? timer;
 
   String SDKKey = 'jJbXtCPPsbt3KpaPMqkCgL7t0zWoHwsKb98k';
   String SDKSecret = 'qmo84JZAg24dguzikPZX49QwajdhpxBiYtu3';
 
-  MeetingWidget({Key key, meetingId, meetingPassword}) : super(key: key) {
+  MeetingWidget({Key? key, meetingId, meetingPassword}) : super(key: key) {
     this.zoomOptions = new ZoomOptions(
       domain: "zoom.us",
       appKey: SDKKey,
@@ -32,14 +30,20 @@ class MeetingWidget extends StatelessWidget {
         disableShare: "true",
         noAudio: "false",
         noDisconnectAudio: "false",
-    );
+        displayName: '',
+        noMore: '',
+        noParticipant: '',
+        noTitle: '',
+        zoomAccessToken: '',
+        zoomToken: '');
   }
 
   bool _isMeetingEnded(String status) {
     var result = false;
 
     if (Platform.isAndroid)
-      result = status == "MEETING_STATUS_DISCONNECTING" || status == "MEETING_STATUS_FAILED";
+      result = status == "MEETING_STATUS_DISCONNECTING" ||
+          status == "MEETING_STATUS_FAILED";
     else
       result = status == "MEETING_STATUS_IDLE";
 
@@ -51,53 +55,51 @@ class MeetingWidget extends StatelessWidget {
     // Use the Todo to create the UI.
     return Scaffold(
       appBar: AppBar(
-          title: Text('Loading meeting '),
+        title: Text('Loading meeting '),
       ),
       body: Padding(
-        padding: EdgeInsets.all(16.0),
-        child: ZoomView(onViewCreated: (controller) {
+          padding: EdgeInsets.all(16.0),
+          child: ZoomView(onViewCreated: (controller) {
+            print("Created the view");
 
-          print("Created the view");
+            controller.initZoom(this.zoomOptions!).then((results) {
+              print("initialised");
+              print(results);
 
-          controller.initZoom(this.zoomOptions)
-              .then((results) {
-
-            print("initialised");
-            print(results);
-
-            if(results[0] == 0) {
-
-              controller.zoomStatusEvents.listen((status) {
-                print("Meeting Status Stream: " + status[0] + " - " + status[1]);
-                if (_isMeetingEnded(status[0])) {
-                  Navigator.pop(context);
-                  timer?.cancel();
-                }
-              });
-
-              print("listen on event channel");
-
-              controller.joinMeeting(this.meetingOptions)
-                  .then((joinMeetingResult) {
-
-                timer = Timer.periodic(new Duration(seconds: 2), (timer) {
-                  controller.meetingStatus(this.meetingOptions.meetingId)
-                      .then((status) {
-                    print("Meeting Status Polling: " + status[0] + " - " + status[1]);
-                  });
+              if (results[0] == 0) {
+                controller.zoomStatusEvents.listen((status) {
+                  print("Meeting Status Stream: " +
+                      status[0] +
+                      " - " +
+                      status[1]);
+                  if (_isMeetingEnded(status[0])) {
+                    Navigator.pop(context);
+                    timer?.cancel();
+                  }
                 });
 
-              });
-            }
+                print("listen on event channel");
 
-          }).catchError((error) {
-
-            print("Error");
-            print(error);
-          });
-        })
-      ),
+                controller
+                    .joinMeeting(this.meetingOptions!)
+                    .then((joinMeetingResult) {
+                  timer = Timer.periodic(new Duration(seconds: 2), (timer) {
+                    controller
+                        .meetingStatus(this.meetingOptions!.meetingId!)
+                        .then((status) {
+                      print("Meeting Status Polling: " +
+                          status[0] +
+                          " - " +
+                          status[1]);
+                    });
+                  });
+                });
+              }
+            }).catchError((error) {
+              print("Error");
+              print(error);
+            });
+          })),
     );
   }
-
 }
