@@ -1,10 +1,10 @@
 
 class ZoomOptions {
 
-  String domain;
-  String appKey;
-  String appSecret;
-  String jwtToken;
+  String? domain;
+  String? appKey;
+  String? appSecret;
+  String? jwtToken;
 
   ZoomOptions({
     this.domain,
@@ -16,21 +16,21 @@ class ZoomOptions {
 
 class ZoomMeetingOptions {
 
-  String userId;
-  String displayName;
-  String meetingId;
-  String meetingPassword;
-  String zoomToken;
-  String zoomAccessToken;
-  String disableDialIn;
-  String disableDrive;
-  String disableInvite;
-  String disableShare;
-  String noDisconnectAudio;
-  String noAudio;
-  String noTitle;
-  String noParticipant;
-  String noMore;
+  String? userId;
+  String? displayName;
+  String? meetingId;
+  String? meetingPassword;
+  String? zoomToken;
+  String? zoomAccessToken;
+  String? disableDialIn;
+  String? disableDrive;
+  String? disableInvite;
+  String? disableShare;
+  String? noDisconnectAudio;
+  String? noAudio;
+  String? noTitle;
+  String? noParticipant;
+  String? noMore;
 
   ZoomMeetingOptions({
     this.userId,
