@@ -9,15 +9,15 @@ typedef void ZoomViewCreatedCallback(ZoomViewController controller);
 
 class ZoomView extends StatefulWidget {
   const ZoomView({
-    Key key,
+    Key? key,
     this.zoomOptions,
     this.meetingOptions,
     this.onViewCreated,
   }) : super(key: key);
 
-  final ZoomViewCreatedCallback onViewCreated;
-  final ZoomOptions zoomOptions;
-  final ZoomMeetingOptions meetingOptions;
+  final ZoomViewCreatedCallback? onViewCreated;
+  final ZoomOptions? zoomOptions;
+  final ZoomMeetingOptions? meetingOptions;
 
   @override
   State<StatefulWidget> createState() => _ZoomViewState();
@@ -48,7 +48,7 @@ class _ZoomViewState extends State<ZoomView> {
     }
 
     var controller = new ZoomViewController._(id);
-    widget.onViewCreated(controller);
+    widget.onViewCreated!(controller);
   }
 }
 
@@ -62,59 +62,61 @@ class ZoomViewController {
   final MethodChannel _methodChannel;
   final EventChannel _zoomStatusEventChannel;
 
-  Future<List> initZoom(ZoomOptions options) async {
+  Future initZoom(ZoomOptions options) async {
     assert(options != null);
 
     var optionMap = new Map<String, String>();
-    optionMap.putIfAbsent("appKey", () => options.appKey);
-    optionMap.putIfAbsent("appSecret", () => options.appSecret);
-    optionMap.putIfAbsent("domain", () => options.domain);
+    optionMap.putIfAbsent("appKey", () => options.appKey!);
+    optionMap.putIfAbsent("appSecret", () => options.appSecret!);
+    optionMap.putIfAbsent("domain", () => options.domain!);
 
     return _methodChannel.invokeMethod('init', optionMap);
   }
 
-  Future<bool> startMeeting(ZoomMeetingOptions options) async {
+  Future startMeeting(ZoomMeetingOptions options) async {
     assert(options != null);
     var optionMap = new Map<String, String>();
-    optionMap.putIfAbsent("userId", () => options.userId);
-    optionMap.putIfAbsent("displayName", () => options.displayName);
-    optionMap.putIfAbsent("meetingId", () => options.meetingId);
-    optionMap.putIfAbsent("meetingPassword", () => options.meetingPassword);
-    optionMap.putIfAbsent("zoomToken", () => options.zoomToken);
-    optionMap.putIfAbsent("zoomAccessToken", () => options.zoomAccessToken);
-    optionMap.putIfAbsent("disableDialIn", () => options.disableDialIn);
-    optionMap.putIfAbsent("disableDrive", () => options.disableDrive);
-    optionMap.putIfAbsent("disableInvite", () => options.disableInvite);
-    optionMap.putIfAbsent("disableShare", () => options.disableShare);
-    optionMap.putIfAbsent("noDisconnectAudio", () => options.noDisconnectAudio);
-    optionMap.putIfAbsent("noAudio", () => options.noAudio);
-    optionMap.putIfAbsent("noTitle", () => options.noTitle);
-    optionMap.putIfAbsent("noParticipant", () => options.noParticipant);
-    optionMap.putIfAbsent("noMore", () => options.noMore);
+    optionMap.putIfAbsent("userId", () => options.userId!);
+    optionMap.putIfAbsent("displayName", () => options.displayName!);
+    optionMap.putIfAbsent("meetingId", () => options.meetingId!);
+    optionMap.putIfAbsent("meetingPassword", () => options.meetingPassword!);
+    optionMap.putIfAbsent("zoomToken", () => options.zoomToken!);
+    optionMap.putIfAbsent("zoomAccessToken", () => options.zoomAccessToken!);
+    optionMap.putIfAbsent("disableDialIn", () => options.disableDialIn!);
+    optionMap.putIfAbsent("disableDrive", () => options.disableDrive!);
+    optionMap.putIfAbsent("disableInvite", () => options.disableInvite!);
+    optionMap.putIfAbsent("disableShare", () => options.disableShare!);
+    optionMap.putIfAbsent(
+        "noDisconnectAudio", () => options.noDisconnectAudio!);
+    optionMap.putIfAbsent("noAudio", () => options.noAudio!);
+    optionMap.putIfAbsent("noTitle", () => options.noTitle!);
+    optionMap.putIfAbsent("noParticipant", () => options.noParticipant!);
+    optionMap.putIfAbsent("noMore", () => options.noMore!);
 
     return _methodChannel.invokeMethod('start', optionMap);
   }
 
-  Future<bool> joinMeeting(ZoomMeetingOptions options) async {
+  Future joinMeeting(ZoomMeetingOptions options) async {
     assert(options != null);
     var optionMap = new Map<String, String>();
-    optionMap.putIfAbsent("userId", () => options.userId);
-    optionMap.putIfAbsent("meetingId", () => options.meetingId);
-    optionMap.putIfAbsent("meetingPassword", () => options.meetingPassword);
-    optionMap.putIfAbsent("disableDialIn", () => options.disableDialIn);
-    optionMap.putIfAbsent("disableDrive", () => options.disableDrive);
-    optionMap.putIfAbsent("disableInvite", () => options.disableInvite);
-    optionMap.putIfAbsent("disableShare", () => options.disableShare);
-    optionMap.putIfAbsent("noDisconnectAudio", () => options.noDisconnectAudio);
-    optionMap.putIfAbsent("noAudio", () => options.noAudio);
-    optionMap.putIfAbsent("noTitle", () => options.noTitle);
-    optionMap.putIfAbsent("noParticipant", () => options.noParticipant);
-    optionMap.putIfAbsent("noMore", () => options.noMore);
+    optionMap.putIfAbsent("userId", () => options.userId!);
+    optionMap.putIfAbsent("meetingId", () => options.meetingId!);
+    optionMap.putIfAbsent("meetingPassword", () => options.meetingPassword!);
+    optionMap.putIfAbsent("disableDialIn", () => options.disableDialIn!);
+    optionMap.putIfAbsent("disableDrive", () => options.disableDrive!);
+    optionMap.putIfAbsent("disableInvite", () => options.disableInvite!);
+    optionMap.putIfAbsent("disableShare", () => options.disableShare!);
+    optionMap.putIfAbsent(
+        "noDisconnectAudio", () => options.noDisconnectAudio!);
+    optionMap.putIfAbsent("noAudio", () => options.noAudio!);
+    optionMap.putIfAbsent("noTitle", () => options.noTitle!);
+    optionMap.putIfAbsent("noParticipant", () => options.noParticipant!);
+    optionMap.putIfAbsent("noMore", () => options.noMore!);
 
     return _methodChannel.invokeMethod('join', optionMap);
   }
 
-  Future<List> meetingStatus(String meetingId) async {
+  Future meetingStatus(String meetingId) async {
     assert(meetingId != null);
 
     var optionMap = new Map<String, String>();
